==> research/02_data_validation.ipynb <==
import os

%pwd

os.chdir("../")
%pwd

import pandas as pd
df = pd.read_csv("./artifacts/data_ingestion/blooms_dataset.csv")
df.info()

from pathlib import Path 
from dataclasses import dataclass

@dataclass(frozen=True)
class DataValidationConfig:
    root_dir: Path
    unzip_data_path: Path
    STATUS_FILE: str 
    all_schema: dict

from blooms.utils.common import read_yaml,create_directories
from blooms.constants import *


class ConfigurationManager:
    def __init__(
        self,
        config_filepath=CONFIG_FILE_PATH,
        params_filepath=PARAMS_FILE_PATH,
        schema_filepath=SCHEMA_FILE_PATH,
    ):
        self.config = read_yaml(config_filepath)
        self.params = read_yaml(params_filepath)
        self.schema = read_yaml(schema_filepath)

        create_directories([self.config.artifacts_root])

    def get_data_validation_config(self) -> DataValidationConfig:
        config = self.config.data_validation
        schema = self.schema.COLUMNS

        create_directories([config.root_dir])

        data_validation_config = DataValidationConfig(
            STATUS_FILE=config.STATUS_FILE,
            root_dir=config.root_dir,
            unzip_data_path=config.unzip_data_path,
            all_schema=schema,
        )
        return data_validation_config

from blooms import logger 
import os 
import pandas as pd

class DataValidation:
    def __init__(self,config:DataValidationConfig):
        self.config = config

    def validate_all_columns(self) -> bool:
        try: 
            validation_status = None 
            data = pd.read_csv(self.config.unzip_data_path)
            all_cols = list(data.columns)
            all_schema = self.config.all_schema.keys()
            for col in all_cols:
                if col not in all_schema:
                    validation_status = False 
                    with open(self.config.STATUS_FILE,'w') as f:
                        f.write(f"Validation Status: {validation_status}")
                else:
                    validation_status = True 
                    with open(self.config.STATUS_FILE,'w') as f:
                        f.write(f"Validattion Status: {validation_status}")

                return validation_status
        except Exception as e:
            raise e

try: 
    config = ConfigurationManager()
    data_validation_config = config.get_data_validation_config()
    data_validation = DataValidation(config=data_validation_config)
    data_validation.validate_all_columns()
except Exception as e:
    raise e
